Front: improve front filters JS customisation
Add the target of the scrolling after product list load into a global variable. This can be changed by 3rd party to control the scroll enabling it to even be disabled.

Add a custom event to warn when the product list is about to be loaded

Use jQuery .get() instead of .load() in ajax_content in order to be able to replace the DOM element with the content loaded

Move debounce to utils file and expose it to a global object

Refs MYS-143
