+++ shuup/front/static_src/js/utils.js
@@ -1,0 +1,32 @@
+/**
+ * This file is part of Shuup.
+ *
+ * Copyright (c) 2012-2019, Shoop Commerce Ltd. All rights reserved.
+ *
+ * This source code is licensed under the OSL-3.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+// https://davidwalsh.name/javascript-debounce-function
+export const debounce = function(func, wait, immediate) {
+    let timeout;
+    return function () {
+        const context = this, args = arguments;
+        const later = function () {
+            timeout = null;
+            if (!immediate) {
+                func.apply(context, args);
+            }
+        };
+        const callNow = immediate && !timeout;
+        clearTimeout(timeout);
+        timeout = setTimeout(later, wait);
+        if (callNow) {
+            func.apply(context, args);
+        }
+    };
+};
+// expose to the world
+window.ShuupToolbox = {
+    debounce
+};
